refactor: remove payment-paypal.py and added to app.py
